fix: update protogalaxy cmake dependencies (#5066)
Closes https://github.com/AztecProtocol/barretenberg/issues/854.

We previously had to make ultra_honk dependent on protogalaxy because we
had the ultra composer in the ultra_honk folder, and that composer had
protogalaxy related functions. Now, with the composer dead, we can have
the correct dependency, which is going the other way.

Old graph:

![barretenberg_modules](https://github.com/AztecProtocol/aztec-packages/assets/42556996/99866e8d-eaa8-45de-855e-d2c886c4f0a0)

New graph:

![barretenberg_modules](https://github.com/AztecProtocol/aztec-packages/assets/42556996/210301b2-abb9-469b-a6b3-c50832f318fa)

No benchmarking update since this should not affect it.