feat(Authwit): lookup the validity of authwits (#5316)
Fixes #4822.

The current solution is pretty ugly (I think).

The issues encountered is mainly that:
- There are no context so we need to pass stuff that seems quite
redundant (address of the contract itself and block number)
- Fetching witness data with oracles explodes if you try to fetch
something that does not exists.
- The implementation here is a lookup from "outside" a transaction.

The latter is the main trouble. It can be somewhat mitigated by having a
function on the typescript side of things that try doing the lookup in
its local storage first to not encounter that case, but then we need to
be able to split the check so seems a bit meh to fix it that way.

This means that the unconstrained cannot really be used nicely as what
@spalladino did at the offsite, but mainly is an off-chain thing for the
user who is unsure if something was approved and want to check it.

I'm not really sure if I like to have the implementation this way where
we are doing things in both ts and noir to get it somewhat working, but
seems like the thing that is making it the easiest to get a query going
fast.