feat(avm): Set gas allowance in public calls (#5567)
Allows user to set gas allowance in public calls, defaulting to the gas
left in the current frame.

- Adds the `(L1|L2|DA)GASLEFT` instructions.
- Adds a new `GasOpts` option to the
`(Public|Avm)Context.call_public_function` so the user can specify the
gas allocation for a nested public call (ignored in public context, only
used in avm context). Defaults gas to the result of gasleft, minus a
buffer for L2 compute.
- Updates the noir interface codegen to handle this new argument.