feat!: trap with revert data (#5732)
This PR adds revert data for user-defined messages in the TRAP opcode.
Not all messages are returned as revert data, since compiler intrinsic
messages are spammed all over SSA and codegening revert data for all
those makes brillig function size blow up and impacts performance in a
hard way due to deserializing load.
This is currently only for static assert messages, since dynamic ones
are implemented as an oracle and probably need a rework to be able to
return them as revert data.

---------

Co-authored-by: Maxim Vezenov <mvezenov@gmail.com>