feat: Typings generator with generics (#6235)
This PR adds a demonomorphizer in the `generate_ts_from_abi` script/tool
so we can use generic types in the protocol circuits interface. It also
fixes the MembershipWitness issue where we had to use concrete types
instead of a generic due to the old limitation of the types generator.