chore: Do not rebuild yarn-projects on bench-comment (#6396)
The bench-comment job was re-running bootstrap.sh on yarn-project (see
[here](https://github.com/AztecProtocol/aztec-packages/actions/runs/9081368520/job/24957398753?pr=6211#step:7:314),
most likely because the "Download base benchmark" step on ci.yml was
downloading a new file to `yarn-projects`, which caused the `COPY . .`
in yarn project build to miss cache.

This downloads the bench files outside the source folder, so it doesn't
trigger a rebuild.