feat: Claim and pay fee on the same tx (#6579)
Allows claiming bridged gas token funds from L1 and use them for paying
fees on the same tx. This required two changes, as described in #6562:

> - In the base rollup circuit, if there's already a `public_data_write`
for the user's balance, the circuit should upate that entry to discount
the fee, instead of inserting a new `protocol_public_data_write`. This
requires updating the `block_building_helpers` and the
`public_processor` on ts-land accordingly.
> - In the sequencer, during tx validation, we should look for public
function calls that trigger a `claim`, and consider that balance when
checking whether the user's balance is enough to pay for the gas limit
at max prices. To avoid having to publicly simulate `claim_public`, we
should instead add a `claim_private` function to the `GasToken` that
triggers an internal public `increase_balance`, and have the sequencer
look for _that_ call, that this needs to be executed during `setup`.

Implementing the above required replacing the `protocol_data_updates`
(which included the data update request for the fee updated generated by
the protocol) with a `final_data_updates` (which is equal to the user
data update requests mixed in with the fee update, since the fee update
can now be a new entry or modify an existing one).

Fixes #6562