feat: new test program for verifying honk (#6781)
Adds a new test Noir program verify_honk_proof that takes in a honk
proof and calls the honk recursive verifier.

Fixes a couple of errors in the flow: we shouldn't be doing vkey hash stuff when processing the vkey anymore since taht isn't being attached, and we need to be able to figure out the proof size and public input size without having valid witnesses (for the write_vk_honk flow).