feat!: make note_getter return BoundedVec instead of an Option array (#7050)
`note_getter` was collapsing all returned notes at the beginning of the
options array, which allowed us to iterate over `options.limit` instead
of `len()`, but this was quite tricky to realize and harder to explain
and realize was correct. This new interface avoids this issue entirely,
though it does require for the user to be familiar with usage of
`BoundedVec` and the fact that Noir loops must be compile-time bounded.

This slightly reduces gate counts since we no longer have all of the
`_is_some` booleans, but the big win will be once we begin setting
`limit` properly. Also, once this is merged, we can use the
`utils::colapse` function from #7016 inside `note_getter` for further
gate count reductions.