feat: Constant Honk proof sizes (#6954)
Introduces logic in the Honk provers/verifiers to 1) produce a proof of
fixed size regardless of the size of the circuit, and 2) produce a
recursive verifier circuit that is independent of the size of the
circuit whose proof it is verifying. This was done largely for
convenience so that no additional logic is needed downstream of bberg to
handle the fact that Honk proofs and recursive verifier circuits are in
general dependent on the size of the inner circuit (unlike for Plonk).

---------

Co-authored-by: lucasxia01 <lucasxia01@gmail.com>
Co-authored-by: ledwards2225 <l.edwards.d@gmail.com>