feat!: extend storage read oracle to receive address and block number (#7243)
Follow up to #7237, closes
https://github.com/AztecProtocol/aztec-packages/issues/7230. I only
changed the oracles and not the PXE interface to keep this change as
small as possible.

I did change the node interface, but made it so you can still do it the
old way by passing `'latest'`, which I had to do in a couple places.

Finally, I added getters for `UnconstrainedContext`, mirroring the work
in #7320, which I imagine are the ones we'll use in the vast majority of
cases.