fix: included argshash computation in public call_interfaces and cleanup (#7354)
During the development of the defunct
https://github.com/AztecProtocol/aztec-packages/pull/6751 I realized we
weren't constraining the args_hash computation in public interfaces, but
promptly forgot during the offsite craze.

This PR fixes that and cleanups the interface computation for
readability. I would have loved to include the checks in the call
interface itself instead of having it in the less clear macros, but we
remove an `if` statement if the function takes no args, which in circuit
land is very good.