feat: enforce limits for each side effect type in AVM (#8889)
The AVM needs to ensure that TX-wide limits for side effects are never
surpassed. This is because if these limits are surpassed, the kernel
cannot prove.

Also, add `contractAddress` to tracing for `writeL2ToL1Message`.

Caveats/todos:
- Does not take into account parent/previous-kernel lengths (handled in
another PR in stack).
- In parent trace, does not take into account side effects created in
nested call/trace. That will be handled when trace operates at the
enqueued-call level (see draft PR at top of stack).