feat: biggroup_goblin handles points at infinity + 1.8x reduction in ECCVM size (#9366)
This PR adds support for biggroup_goblin handling elliptic curve points
at infinity

This feature is used to optimize the ECCVM instructions created when
running a recursive protogalaxy verifier.

Instead of performing size-2 scalar multiplications for every
witness/commitment, additional random challenges are generated in order
to evaluate two large batch multiplications.

The technique implemented is described in
https://hackmd.io/T1239dufTgO1v8Ie7EExlQ?view

In the case of ClientIVCRecursionTests.ClientTubeBase where the number
of circuits is set to four, this reduces the size of the ECCVM execution
trace size by 45%, which is good enough to reduce the log dyadic size of
the ClientIVCVerifier by 1/2.