feat: IPA Accumulation implementation (#9494)
Adds new functions to the IPA class that accumulate two IPA claims in
circuit, and does prover work to prove new accumulated IPA claim.

Also adds new tests for IPA recursion. The new RecursiveSmall test
allowed us to find a subtle bug in IPA where one value was not being
negated, leading to an incorrect identity. This was not caught
previously because this value, the evaluation, was always 0 in the
existing tests, as Shplonk always sets it to 0.