fix: deploy l2 contracts fails on 48 validator (#9860)
This was sometimes erroring out on the first attempt while awaiting the
transaction to settle,
but then subsequent calls were attempting to reinitialize.

Added better logging to help clarify.

Also adjust the validator url in the template to be the top level
service, to better distribute transactions across the network and not
rely so heavily on the boot node for gossiping.

---------

Co-authored-by: ludamad <adam.domurad@gmail.com>