chore: add tests for gov proposer (#11633)
Key changes:

- Add an e2e and spartan test for proposing a new payload to the
governance proposer, and executing it (i.e. passing it up to
governance); follow-on work is needed to test execution by governance.
- Use `forge` to generate an artifact for the rollup contract's storage,
which we copy into the l1-artifacts project; use this artifact to when
overriding the slot for blob checks.
- The typescript helper for the forwarder contract now explicitly takes
a ref to the rollup in use to aid in error extraction.
- Improvement made in extracting nested error codes from forward
functions

Fix #11681