feat(avm): drop bytecode-specific hints (#12234)
This PR removes the `AvmBytecodeHints` and breaks them up into lower level hints. Two of them already existed (ContractInstances and NullifierChecks) but I had to add a new one for ContractClass.

I'm keeping `getBytecode` in the journal for now, but it is now calling other methods _in the journal_, instead of using the DB directly. Therefore, it is indirectly generating hints and tracing things as well. This let us get rid of duplicated code for getting instances (and update hints and membership check hints).

No shims in this PR, just painful refactoring :)