chore: remove some unnecessary mod.nr files (#12797)
This is ultimately a styling choice, I find it annoying that the code
for both `note_getter` and most of the macros is in `mod.nr` files,
which make it a bit harder to find the content, and make it quite
confusing when you have five of those files open in the IDE and you
can't tell them apart by name. Ideally if we have a single mod `foo`
we'd just have `foo.nr`, and only do `foo/mod.nr` once there's multiple
files, keeping the `mod` as a source of further mode declarations and
potentially exports, but not actual source code.