feat: Zw/goblin avm tests (#12904)
Introduces a new class `AvmGoblinRecursiveVerifier` that manages the
Goblinized AVM recursive verification algorithm. The algorithm involves
the construction of two circuits. The first is a Mega-arithmetized AVM
recursive verifier. The second is an Ultra circuit that recursively
verifies the proof of the first (which consists of a Mega proof plus a
Goblin proof). The output of the algorithm is equivalent to a that of a
Rollup Honk recursive verifier: a pairing point accumulator plus an IPA
claim/proof (stemming from recursive verification of the ECCVM). See the
class description for a more detailed description of the algorithm.

Currently the algorithm is tested in a standalone test only and is not
yet integrated into ACIR. E.g. at this stage a noir program with a
verify_proof call of type AVM will still generate a vanilla Ultra
recursive verifier.

Note: This work is incomplete in a number of ways, the most significant
of which I've documented in several sub-issues under this umbrella
[issue](https://github.com/AztecProtocol/barretenberg/issues/1303).

---------

Co-authored-by: zac-williamson <blorktronics@gmail.com>