feat(avm): checkpointing hints (#13302)
The hints are fully implemented but it is NOT being extensively used in
the code. I only added one call to `create_checkpoint` in the
`tx_execution` but the use in TS is complex and intertwined with
insertion/call success/failure and I didn't want to understand and
implement all that in this PR. This has to be implemented once the
tx_execution is implemented.