test: `e2e_state_vars ` not using `DocsExampleContract` (#13368)
This PR is part of a series of PRs in which I clean up our use of test
contracts. In this case I am replacing the use of DocsExampleContract in
e2e_state_vars.test.ts with the goal of eventually not using
DocsExampleContract in /yarn-project at all.

I introduce a new StateVars contract.