feat!: rename encrypted_logs to messages (#13496)
Further renaming and moving incoming, but this is the largest piece as
it affects all files, so having this as a separate part of the changeset
will avoid very messy PR diffs.

edit: ended up making a much larger changeset, since it's all renamings.
Summary:

```
- `encrypted_logs` to `messages`: this module now handles much more than just encrypted logs (including unconstrained message delivery, message encoding, etc.)
- `log_assembly_strategies` to `logs`
- `discovery` moved to `messages`: given that what is discovered are messages
- `default_aes128` removed
```

Which means there's no longer an `encrypted_logs` directory in
`aztec-nr` (which I always found odd) and we instead have `messages`,
with the following content:

```
$ tree messages/
messages/
├── discovery
│   ├── mod.nr
│   ├── nonce_discovery.nr
│   ├── partial_notes.nr
│   ├── pending_tagged_log.nr
│   ├── private_logs.nr
│   └── private_notes.nr
├── encoding.nr
├── encryption
│   ├── aes128.nr
│   ├── log_encryption.nr
│   ├── mod.nr
│   └── poseidon2.nr
├── logs
│   ├── arithmetic_generics_utils.nr
│   ├── event.nr
│   ├── mod.nr
│   ├── note.nr
│   └── utils.nr
├── mod.nr
└── msg_type.nr
```

which seems fairly reasonable: discovery is about discovering messages,
encoding is of messages, encryption is arguably not strictly just
messages but it _is_ the only usage we have for it, and logs is about
how to put a message in a log.